controller:  move builds.go => controller/builds
